% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accepted_taxa}
\alias{accepted_taxa}
\title{Accepted Taxa}
\format{
A data frame with 23594 rows and 2 columns, the definitions of which are:
\describe{
\item{TVK}{The UKSI taxon version keys for the accepted taxa.}
\item{taxon_name}{The taxon names of the accepted taxa.}
}
}
\usage{
accepted_taxa
}
\description{
The taxon names accepted by RMAVIS. Equivalent to the unique species in the \code{UKVegTB::taxonomic_backbone} 'full_name' column
with the addition of taxa with strata suffixes as present in \code{RMAVIS::nvc_taxa_lookup}.
}
\details{
\code{accepted_taxa}
}
\keyword{datasets}
