% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_taxa.R
\name{aggregate_taxa}
\alias{aggregate_taxa}
\title{Combine taxa into aggregated groups}
\usage{
aggregate_taxa(
  plot_data,
  agg_lookup,
  plot_data_taxon_col,
  agg_lookup_taxon_col
)
}
\arguments{
\item{plot_data}{A data frame containing vegetation plot survey data with 5 columns: Year, Group, Quadrat, Species, and Cover. See \code{MNNPC::mnnpc_example_data}.}

\item{agg_lookup}{A data frame containing a lookup between accepted taxon names present in plot_data and taxa or groups of taxa combined for analysis. See \code{MNNPC::mnnpc_taxa_conv}.}

\item{plot_data_taxon_col}{The name of the column in plot_data containing the species/taxon names.}

\item{agg_lookup_taxon_col}{The name of the column in agg_lookup containing the species/taxon names.}
}
\value{
A data frame containing the aggregated data
}
\description{
Combine taxa into aggregated groups by summing the cover using a lookup.
}
\examples{
RMAVIS::aggregate_taxa(plot_data = MNNPC::mnnpc_example_data$`St. Croix State Forest`, 
                       agg_lookup = MNNPC::mnnpc_taxa_conv,
                       plot_data_taxon_col = "Species",
                       agg_lookup_taxon_col = "taxon")
}
