% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_habCor_final}
\alias{all_habCor_final}
\title{All habitat correspondence data}
\format{
A data frame with 7641 rows and 5 columns, the definitions of which are:
\describe{
\item{NVC.Code}{The NVC community code.}
\item{Relationship}{The relationship between the NVC community and the habitat present in the 'label' column.}
\item{Code}{The code associated with the habitat present in the 'label' column.}
\item{Label}{The corresponding habitat.}
\item{Classification}{The habitat classification system.}
}
}
\usage{
all_habCor_final
}
\description{
A dataset containing habitat correspondences between NVC communities and the
following classifications.
\itemize{
\item UKHab - Level 2
\item UKHab - Level 3
\item UKHab - Level 4
\item UKHab - Level 5
\item Biodiversity Action Plan Priority Habitats
\item Phase 1 Habitat Classification
\item Biodiversity Action Plan Broad Habitats
\item EUNIS Classification
\item Vegetation communities of British lakes
\item EU Habitats Directive Annex 1
\item NPMS Broad
\item NPMS Fine
}
}
\details{
\code{all_habCor_final}
}
\keyword{datasets}
