% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{broad_habitat_indicators}
\alias{broad_habitat_indicators}
\title{SOWG floristic tables}
\format{
A data frame with 1884 rows and 23 columns, the definitions of which are:
\describe{
\item{recommended_taxon_name}{The taxon name, see \code{UKVegTB::taxonomic_backbone}}
\item{recommended_TVK}{The UKSI taxon version key, see \code{UKVegTB::taxonomic_backbone}.}
\item{1}{The broad habitat 'Broadleaved, mixed and yew woodland'.}
\item{2}{The broad habitat 'Coniferous woodland'.}
\item{3}{The broad habitat 'Boundary and linear features'.}
\item{4}{The broad habitat 'Arable and horticultural'.}
\item{5}{The broad habitat 'Improved grassland'.}
\item{6}{The broad habitat 'Neutral grassland'.}
\item{7}{The broad habitat 'Calcareous grassland'.}
\item{8}{The broad habitat 'Acid grassland'.}
\item{9}{The broad habitat 'Bracken'.}
\item{10}{The broad habitat 'Dwarf shrub heath'.}
\item{11}{The broad habitat 'Fen, marsh and swamp'.}
\item{12}{The broad habitat 'Bog'.}
\item{13}{The broad habitat 'Standing water and canals'.}
\item{14}{The broad habitat 'Rivers and streams'.}
\item{15}{The broad habitat 'Montane habitats'.}
\item{16}{The broad habitat 'Inland rock'.}
\item{17}{The broad habitat 'Built-up areas and gardens'.}
\item{18}{The broad habitat 'Supralittoral rock'.}
\item{19}{The broad habitat 'Supralittoral sediment'.}
\item{21}{The broad habitat 'Littoral sediment'.}
\item{23}{The broad habitat 'Inshore sublittoral sediment'.}
}
}
\usage{
broad_habitat_indicators
}
\description{
A dataset containing the floristic tables for each Scottish Oceanic Wet Grassland NVC community.
}
\details{
\code{sowg_floristic_tables}
}
\keyword{datasets}
