% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{call_nbn_bulk_lookup}
\alias{call_nbn_bulk_lookup}
\title{Use NBN Bulk Species Lookup API}
\usage{
call_nbn_bulk_lookup(species)
}
\arguments{
\item{species}{}
}
\description{
Uses the NBN Bulk Species Lookup API (https://api.nbnatlas.org/#ws87) to
retrieve taxonomic information on a vector or list of species names.
}
