% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{calthion_cm_he}
\alias{calthion_cm_he}
\title{Calthion community-mean Hill-Ellenberg values}
\format{
A data frame with 95 rows and 4 columns, the definitions of which are:
\describe{
\item{nvc_code}{The Calthion unit code.}
\item{indicator}{The Hill-Ellenberg indicator code, one of: F, L, N, R, S}
\item{mean}{The mean indicator value.}
\item{sd}{The standard deviation of the indicator value.}
}
}
\usage{
calthion_cm_he
}
\description{
A dataset containing community-mean Hill-Ellenberg environmental indicator valueas for each Calthion community.
}
\details{
\code{calthion_cm_he}
}
\keyword{datasets}
