% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{calthion_pquads}
\alias{calthion_pquads}
\title{Calthion pseudo-quadrats}
\format{
A data frame with 9408 rows and 3 columns, the definitions of which are:
\describe{
\item{nvc_code}{The calthion unit code.}
\item{psq_id}{The pseudo-quadrat ID}
\item{nvc_taxon_name}{The taxon name, see \code{RMAVIS::nvc_taxa_lookup}.}
}
}
\usage{
calthion_pquads
}
\description{
A dataset containing pseudo-quadrats for each Calthion community unit.
}
\details{
\code{calthion_pquads}
}
\keyword{datasets}
