% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{concordance}
\alias{concordance}
\title{Concordance information for taxon names present in RMAVIS}
\format{
A data frame with 11281 rows and 6 columns, the definitions of which are:
\describe{
\item{bsbiTaxonId}{The BSBI taxon ID}
\item{TVK}{The United Kingdom Species Inventory (UKSI) Taxon Version Key (TVK)}
\item{ddb_name}{The taxon name as present in the BSBI database.}
\item{originalTaxonName}{The taxon name as present in the original NVC volumes.}
\item{rmavisTaxonName}{The taxon name as present in RMAVIS}
\item{TaxonGroup}{The name of the taxonomic group that species belongs to.}
}
}
\usage{
concordance
}
\description{
A dataset containing concordance information to facilitate the matching of
species present in the original NVC dataset (and in the {assignNVC} R package)
to additional taxonomic backbones.
}
\details{
\code{concordance}
}
\keyword{datasets}
