% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{concordance_all}
\alias{concordance_all}
\title{A dataset containing}
\format{
A data frame with 16500 rows and 7 columns, the definitions of which are:
\describe{
\item{}{}
}
}
\usage{
concordance_all
}
\description{
\code{}
}
\keyword{datasets}
