% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{concordance_bryophytes}
\alias{concordance_bryophytes}
\title{A dataset containing}
\format{
A data frame with 1569 rows and 5 columns, the definitions of which are:
\describe{
\item{}{}
}
}
\usage{
concordance_bryophytes
}
\description{
\code{}
}
\keyword{datasets}
