% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable_functions.R
\name{dataListFilter}
\alias{dataListFilter}
\title{Modify a reactable table to provide a dropdown menu}
\usage{
dataListFilter(tableId, style = "width: 100\%; height: 28px;")
}
\arguments{
\item{tableId}{The ID of the table, usually supplied via the 'elementId' of the  \code{reactable::reactable()} function.}
}
\value{
A modified \code{reactable} table with a dropdown
}
\description{
Modify a \code{reactable} table to provide a dropdown mene containing all unique values in a column.
Apply this function in a column definition within the 'columns' or 'defaultColDef' argument of \code{reactable::reactable()} using \code{reactable::colDef()}.
}
\details{
Taken from the \code{reactable} documentation found here: https://glin.github.io/reactable/articles/custom-filtering.html
}
