% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example vegetation survey data}
\format{
A list with 4 entries:
\describe{
\item{Parsonage Down}{}
\item{Leith Hill Place Wood}{}
\item{Whitwell Common}{}
\item{Newborough Warren}{}
}
}
\usage{
example_data
}
\description{
A list of data frames containing four example vegetation survey datasets.
}
\details{
\code{example_data}
}
\keyword{datasets}
