% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data_all}
\alias{example_data_all}
\title{A dataset containing}
\format{
A data frame with 11143 rows and 6 columns, the definitions of which are:
\describe{
\item{}{}
}
}
\usage{
example_data_all
}
\description{
\code{}
}
\keyword{datasets}
