% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{habCor_classifications}
\alias{habCor_classifications}
\title{Habitat Classifications}
\format{
An object of class \code{character} of length 12.
}
\usage{
habCor_classifications
}
\description{
A vector containing the names of the habitat classification systems for which there is correspondence data in \code{RMAVIS::habCor_data}
}
\details{
\code{habCor_classifications}
}
\keyword{datasets}
