% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{habitatRestrictionPrefixes}
\alias{habitatRestrictionPrefixes}
\title{A named list of prefixes for NVC habitats}
\format{
A named list with  entries.
}
\usage{
habitatRestrictionPrefixes
}
\description{
A named list of prefixes for NVC habitats, for each of the broad NVC habitats
(W, M, H, MG, CG, U, A, S, SM, SD, MC, OV, SOWG)
}
\details{
\code{habitatRestrictionPrefixes}
}
\keyword{datasets}
