% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_broad_habitat.R
\name{matchBroadHabitats}
\alias{matchBroadHabitats}
\title{Match species lists to Broad habitats}
\usage{
matchBroadHabitats(df, tvk_col = "recommended_TVK")
}
\arguments{
\item{df}{A data frame including species presences.}

\item{tvk_col}{The column in the dataframe containing the UKSI Taxon Version Keys (TVKs), see \code{UKVegTB::taxonomic_backbone}.}
}
\value{
A data frame containing the top-fitting broad habitats.
}
\description{
Match species lists to Broad habitats
}
\examples{
data <- RMAVIS::example_data$`Parsonage Down` |> 
  dplyr::filter(Quadrat == "3N.1") |> 
  dplyr::left_join(UKVegTB::taxa_lookup[, c("taxon_name", "recommended_TVK")], by = c("Species" = "taxon_name"))
RMAVIS::matchBroadHabitats(df = data, tvk_col = "recommended_TVK")
}
\references{
Jackson, D.L., 2000. Guidance on the interpretation of the Biodiversity Broad Habitat Classification (terrestrial and freshwater types): Definitions and the relationship with other classifications (JNCC Report No. 307). JNCC, Peterborough.
}
