% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nvc_community_attributes}
\alias{nvc_community_attributes}
\title{NVC Community Attributes}
\format{
A data frame with 875 rows and 11 columns, the definitions of which are:
\describe{
\item{fullname}{The full name of the NVC unit.}
\item{name}{The abridged name of the NVC unit.}
\item{nvc_code}{The NVC unit code.}
\item{nvc_code_parent}{The NVC unit code of the parent community, e.g. A2 for A2a.}
\item{basal}{A boolean. If TRUE the NVC unit does not have any child units e.g. A1. If FALSE the NVC unit has child units, e.g. A.}
\item{rank}{The rank of the NVC unit.}
\item{num_samples}{The number of plots (samples) used to define the NVC unit.}
\item{min_species}{The minimum number of species in the plots constituting the NVC unit.}
\item{max_species}{The maximum number of species in the plots constituting the NVC unit.}
\item{mean_species}{The mean number of species in the plots constituting the NVC unit.}
\item{species_count}{The total number of unique species in the plots constituting the NVC unit. Note that this does not include 'rare' species, species that occurred in less than 5\% of plots.}
}
}
\usage{
nvc_community_attributes
}
\description{
Selected attributes for each NVC unit, exported from the original NVC MS Access database.
}
\details{
\code{nvc_community_attributes}
}
\keyword{datasets}
