% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nvc_community_codes}
\alias{nvc_community_codes}
\title{A vector of NVC community codes}
\format{
A vector with 883 entries.
}
\usage{
nvc_community_codes
}
\description{
A vector of NVC community codes
}
\keyword{datasets}
