% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nvc_floristic_tables_comparison}
\alias{nvc_floristic_tables_comparison}
\title{NVC floristic tables comparison}
\format{
A data frame with 46575 rows and 3 columns, the definitions of which are:
\describe{
\item{nvc_code}{The NVC unit code.}
\item{original_name}{The taxon name as it appears in the original NVC MS access database and volumes.}
\item{nvc_taxon_name}{The updated taxon name, see \code{RMAVIS::nvc_taxa_lookup}.}
}
}
\usage{
nvc_floristic_tables_comparison
}
\description{
A lookup table between the original NVC floristic table taxon names and updated taxon names.#'
}
\details{
\code{nvc_floristic_tables_comparison}
}
\keyword{datasets}
