% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nvc_pquad_dca}
\alias{nvc_pquad_dca}
\title{The DCA results for all pseudo-quadrats}
\format{
An object of class "decorana". See \code{vegan::decanora}.
}
\usage{
nvc_pquad_dca
}
\description{
The results of a Detrended Correspondence Analysis (DCA) performed on all
pseudo-quadrats present in \code{RMAVIS::nvc_pquads_wide}.
}
\details{
\code{nvc_pquad_dca}
}
\keyword{datasets}
