% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nvc_pquad_dca_all_centroids}
\alias{nvc_pquad_dca_all_centroids}
\title{A dataset containing}
\format{
A data frame with 888 rows and 5 columns, the definitions of which are:
\describe{
\item{}{}
}
}
\usage{
nvc_pquad_dca_all_centroids
}
\description{
\code{}
}
\keyword{datasets}
