% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nvc_pquad_dca_noBryophytes}
\alias{nvc_pquad_dca_noBryophytes}
\title{A dataset containing}
\format{
A data frame with  rows and  columns, the definitions of which are:
\describe{
\item{}{}
}
}
\usage{
nvc_pquad_dca_noBryophytes
}
\description{
\code{}
}
\keyword{datasets}
