% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nvc_pquads}
\alias{nvc_pquads}
\title{The pseudo-quadrats for each NVC unit}
\format{
A data frame with 981243 rows and 3 columns, the definitions of which are:
\describe{
\item{NVC}{The NVC community or sub-community code}
\item{Pid3}{The pseudo-quadrat ID}
\item{species}{The taxon name}
}
}
\usage{
nvc_pquads
}
\description{
A dataset containing pseudo-quadrats for each NVC unit.
}
\details{
\code{nvc_pquads}
}
\keyword{datasets}
