% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nvc_pquads_final}
\alias{nvc_pquads_final}
\title{A dataset containing}
\format{
A data frame with 981749 rows and 3 columns, the definitions of which are:
\describe{
\item{}{}
}
}
\usage{
nvc_pquads_final
}
\description{
\code{}
}
\keyword{datasets}
