% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nvc_pquads_final_wide}
\alias{nvc_pquads_final_wide}
\title{A dataset containing}
\format{
A data frame with 57391 rows and 1610 columns, the definitions of which are:
\describe{
\item{}{}
}
}
\usage{
nvc_pquads_final_wide
}
\description{
\code{}
}
\keyword{datasets}
