% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nvc_pquads_mean_unweighted_eivs}
\alias{nvc_pquads_mean_unweighted_eivs}
\title{The mean unweighted Hill-Ellenberg values for all pseudo-quadrats}
\format{
A data frame with 57716 rows and 6 columns, the definitions of which are:
\describe{
\item{Pid3}{The pseudo-quadrat ID.}
\item{F}{The mean unweighted Hill-Ellenberg moisture value}
\item{L}{The mean unweighted Hill-Ellenberg light value}
\item{N}{The mean unweighted Hill-Ellenberg nitrogen value}
\item{R}{The mean unweighted Hill-Ellenberg reaction value}
\item{S}{The mean unweighted Hill-Ellenberg salinity value}
}
}
\usage{
nvc_pquads_mean_unweighted_eivs
}
\description{
A data frame containing the mean unweighted Hill-Ellenberg values for all
pseudo-quadrats, calculated using \code{RMAVIS::nvc_pquads} and
\code{RMAVIS::master_data}.
}
\details{
\code{nvc_pquads_mean_unweighted_eivs}
}
\keyword{datasets}
