% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nvc_pquads_mean_unweighted_eivs_noBryophytes}
\alias{nvc_pquads_mean_unweighted_eivs_noBryophytes}
\title{A dataset containing}
\format{
A data frame with 57704 rows and 6 columns, the definitions of which are:
\describe{
\item{}{}
}
}
\usage{
nvc_pquads_mean_unweighted_eivs_noBryophytes
}
\description{
\code{}
}
\keyword{datasets}
