% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nvc_pquads_wide}
\alias{nvc_pquads_wide}
\title{The pseudo-quadrats for each NVC community in wide format}
\format{
A data frame with 57716 rows and 1584 columns.
}
\usage{
nvc_pquads_wide
}
\description{
A data frame containing the pseudo-quadrats for each NVC community in wide
(pseudo-quadrat by species) format
}
\details{
\code{nvc_pquads_wide}
}
\keyword{datasets}
