% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nvc_psquad_cm_he}
\alias{nvc_psquad_cm_he}
\title{NVC pseudo-quadrat community-mean Hill-Ellenberg values}
\format{
A data frame with 21362 rows and 7 columns, the definitions of which are:
\describe{
\item{nvc_code}{The NVC unit code.}
\item{psq_id}{The pseudo-quadrat ID.}
\item{F}{The Hill-Ellenberg Moisture score.}
\item{L}{The Hill-Ellenberg Light score.}
\item{N}{The Hill-Ellenberg Nitrogen score.}
\item{R}{The Hill-Ellenberg Reaction score.}
\item{S}{The Hill-Ellenberg Salinity score.}
}
}
\usage{
nvc_psquad_cm_he
}
\description{
NVC pseudo-quadrat community-mean Hill-Ellenberg values
}
\details{
\code{nvc_psquad_cm_he}
}
\keyword{datasets}
