% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nvc_taxa_lookup}
\alias{nvc_taxa_lookup}
\title{NVC Taxa Lookup}
\format{
A data frame with 1641 rows and 7 columns, the definitions of which are:
\describe{
\item{original_taxon_name}{The taxon name as it appears in the original NVC MS access database and volumes.}
\item{original_TVK}{The UKSI taxon version key associated with the original taxon name, see \code{UKVegTB::taxa_lookup}.}
\item{recommended_taxon_name}{The latest recommended name for the taxon concept recorded as the original name, see \code{UKVegTB::taxonomic_backbone}.}
\item{recommended_TVK}{The UKSI taxon version key associated with the recommended taxon name, see \code{UKVegTB::taxa_lookup}.}
\item{strata}{The strata for the taxon: c = Canopy, S = Shrub, and g = Ground.}
\item{nvc_taxon_name}{The taxon name, equivalent to the recommended taxon names with strata suffixes.}
}
}
\usage{
nvc_taxa_lookup
}
\description{
A lookup between the original taxon names (as present in the NVC MS Access database and volumes)
the associated recommended taxon names in version 20250703a of the UKSI,
and the 'nvc taxon names', which are equivalent to the recommended taxon names with strata suffixes.
}
\details{
\code{nvc_taxa_lookup}
}
\keyword{datasets}
