% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mavis_data.R
\name{read_mavis_data}
\alias{read_mavis_data}
\title{Read an input file downloaded from MAVIS}
\usage{
read_mavis_data(filepath)
}
\arguments{
\item{filepath}{The file path to a .txt file produced by MAVIS.}
}
\value{
A five column data frame containing the MAVIS input data in RMAVIS format.
}
\description{
Read an input file downloaded from the desktop MAVIS application and wrangle
it into the format required by RMAVIS. Namely a data frame with the folowing
columns:
\describe{
\item{Year}{The year, set to the year the function was used.}
\item{Group}{The group, set to "A" by default.}
\item{Quadrat}{The quadrat ID.}
\item{Species}{The species name.}
\item{Cover}{The cover values, note that 0 is replaced with 0.1}
}
}
\examples{
RMAVIS::read_mavis_data(filepath = system.file("testdata", "mavis_testdata.txt", package = "RMAVIS"))
}
