% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{runApp}
\alias{runApp}
\title{Run the \code{RMAVIS} R Shiny app}
\usage{
runApp()
}
\description{
Run the \code{RMAVIS} R Shiny app following the installation of \code{RMAVIS} as a
package.
}
\examples{
RMAVIS::runApp()
}
