% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_functions.R
\name{similarityCzekanowski}
\alias{similarityCzekanowski}
\title{Calculate the similarity between vegetation data using the Czekanowski index}
\usage{
similarityCzekanowski(
  samp_df,
  comp_df,
  samp_species_col,
  comp_species_col,
  samp_group_name,
  comp_group_name,
  samp_weight_name,
  comp_weight_name,
  downweight_threshold = 1,
  downweight_value = 0.1
)
}
\arguments{
\item{samp_df}{A data frame containing either sample vegetation plot data with cover values, or a syntopic/floristic table composed from sample vegetation plot data.}

\item{comp_df}{A data frame containing either}

\item{samp_species_col}{}

\item{comp_species_col}{}

\item{samp_group_name}{}

\item{comp_group_name}{}

\item{samp_weight_name}{}

\item{downweight_threshold}{If 0 no downeighting occurs.}

\item{downweight_value}{}
}
\description{
Calculate the Czekanowski's Quantitative Index using either a set of sample
vegetation plots and a set of reference vegetation plots; or a set of composed
floristic/syntopic tables and a set of references floristic/syntopic tables.
Method follows Bray and Curtis (1957) and  Field and McFarlane (1968).
}
