% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_functions.R
\name{similarityJaccard}
\alias{similarityJaccard}
\title{Title}
\usage{
similarityJaccard(
  samp_df,
  comp_df,
  samp_species_col,
  comp_species_col,
  samp_group_name = "ID",
  comp_group_name = "Pid3",
  comp_groupID_name = "NVC",
  average_comp = TRUE
)
}
\arguments{
\item{samp_df}{}

\item{comp_df}{}

\item{samp_species_col}{}

\item{comp_species_col}{}

\item{samp_group_name}{}

\item{comp_group_name}{}

\item{comp_groupID_name}{}

\item{average_comp}{}
}
\description{
Title
}
