% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sowg_cm_he}
\alias{sowg_cm_he}
\title{SOWG community-mean Hill-Ellenberg values}
\format{
A data frame with 65 rows and 4 columns, the definitions of which are:
\describe{
\item{nvc_code}{The SOWG unit code.}
\item{indicator}{The Hill-Ellenberg indicator code, one of: F, L, N, R, S}
\item{mean}{The mean indicator value.}
\item{sd}{The standard deviation of the indicator value.}
}
}
\usage{
sowg_cm_he
}
\description{
A dataset containing community-mean Hill-Ellenberg environmental indicator valueas for each SOWG community.
}
\details{
\code{sowg_cm_he}
}
\keyword{datasets}
