% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sowg_community_attributes}
\alias{sowg_community_attributes}
\title{SOWG Community Attributes}
\format{
A data frame with 13 rows and 11 columns, the definitions of which are:
\describe{
\item{fullname}{The full name of the NVC unit.}
\item{name}{The abridged name of the NVC unit.}
\item{nvc_code}{The NVC unit code.}
\item{nvc_code_parent}{The NVC unit code of the parent community, e.g. AgBp for AgBpa}
\item{basal}{A boolean. If TRUE the NVC unit does not have any child units e.g. CnPe. If FALSE the NVC unit has child units, e.g. AgBp.}
\item{rank}{The rank of the NVC unit.}
\item{num_samples}{The number of plots (samples) used to define the NVC unit.}
\item{min_species}{The minimum number of species in the plots constituting the NVC unit.}
\item{max_species}{The maximum number of species in the plots constituting the NVC unit.}
\item{mean_species}{The mean number of species in the plots constituting the NVC unit.}
\item{species_count}{The total number of unique species in the plots constituting the NVC unit. Note that this does not include 'rare' species, species that occurred in less than 5\% of plots.}
}
}
\usage{
sowg_community_attributes
}
\description{
Selected attributes for each Scottish Oceanic Wet Grassland NVC unit.
}
\details{
\code{sowg_community_attributes}
}
\keyword{datasets}
