% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sowg_floristic_tables}
\alias{sowg_floristic_tables}
\title{SOWG floristic tables}
\format{
A data frame with 977 rows and 8 columns, the definitions of which are:
\describe{
\item{nvc_code}{The NVC unit code.}
\item{nvc_taxon_name}{The taxon name, see \code{RMAVIS::nvc_taxa_lookup}.}
\item{constancy}{The constancy of occurrence across the plots constituting the NVC unit, see \code{RMAVIS:::constancyConversion}.}
\item{absolute_frequency}{The number of occurrences in the plots constituting the NVC unit.}
\item{relative_frequency}{The proportion of plots constituting the NVC unit that the taxon is present in.}
\item{minimum_cover}{The minimum cover of the taxon in the plots constituting the NVC unit.}
\item{mean_cover}{The mean cover of the taxon in the plots constituting the NVC unit.}
\item{maximum_cover}{The maximum cover of the taxon in the plots constituting the NVC unit.}
}
}
\usage{
sowg_floristic_tables
}
\description{
A dataset containing the floristic tables for each Scottish Oceanic Wet Grassland NVC community.
}
\details{
\code{sowg_floristic_tables}
}
\keyword{datasets}
