% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sowg_pquads}
\alias{sowg_pquads}
\title{SOWG pseudo-quadrats}
\format{
A data frame with 6864 rows and 3 columns, the definitions of which are:
\describe{
\item{nvc_code}{The NVC unit code.}
\item{psq_id}{The pseudo-quadrat ID}
\item{nvc_taxon_name}{The taxon name, see \code{RMAVIS::nvc_taxa_lookup}.}
}
}
\usage{
sowg_pquads
}
\description{
A dataset containing pseudo-quadrats for each Scottish Oceanic Wet Grassland NVC unit.
}
\details{
\code{sowg_pquads}
}
\keyword{datasets}
