% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{speciesNames}
\alias{speciesNames}
\title{A vector of species names accepted in \code{RMAVIS}}
\format{
A vector with 11938 entries.
}
\usage{
speciesNames
}
\description{
A vector of species names accepted in \code{RMAVIS}
}
\keyword{datasets}
