% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{subset_nvcData}
\alias{subset_nvcData}
\title{Subset a data frame of NVC data by habitat}
\usage{
subset_nvcData(nvc_data, habitatRestriction, col_name)
}
\arguments{
\item{nvc_data}{A data frame of pseudo-quadrats.}

\item{habitatRestriction}{A vector of habitat codes, see \code{RMAVIS:::habitatRestriction_options}.}

\item{col_name}{The name of the column containing NVC pseudo-quadrat names with NVC code prefixes.}
}
\value{
A data frame containing the supplied NVC data for selected habitats
}
\description{
Subset a data frame of NVC data such as pseudo-quadrats or syntopic tables by
habitat type by supplying a vector of prefixes.
}
\examples{
RMAVIS::subset_nvcData(nvc_data = RMAVIS::nvc_pquads, habitatRestriction = c("U", "H"), col_name = "psq_id")
}
