% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_vcData.R
\name{subset_vcData}
\alias{subset_vcData}
\title{Subset a data frame of VC data by habitat}
\usage{
subset_vcData(
  vc_data,
  habitatRestriction,
  col_name,
  habitatRestrictionPrefixes = RMAVIS:::habitatRestrictionPrefixes
)
}
\arguments{
\item{habitatRestriction}{A vector of habitat codes, see \code{RMAVIS:::habitatRestrictionPrefixes} or \code{MNNPC::mnnpc_vc_systems_named}.}

\item{col_name}{The name of the column containing VC pseudo-quadrat names with VC code prefixes.}

\item{habitatRestrictionPrefixes}{A named list of vegetation classification types e.g. M in the GB-NVC.}

\item{nvc_data}{A data frame of pseudo-quadrats.}
}
\value{
A data frame containing the supplied VC data for selected habitats
}
\description{
Subset a data frame of VC data such as pseudo-quadrats or syntopic tables by
habitat type by supplying a vector of prefixes.
}
\examples{
RMAVIS::subset_vcData(vc_data = RMAVIS::nvc_pquads, habitatRestriction = c("U", "H"), col_name = "psq_id")
}
