% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxa_lookup}
\alias{taxa_lookup}
\title{Taxa lookup}
\format{
A data frame with 87618 rows and 4 columns, the definitions of which are:
\describe{
\item{taxon_name}{The taxon names of the synonymous taxon concepts associated with the reccomended taxa.}
\item{TVK}{The UKSI taxon version keys associated with the synonymous taxa in taxon_name.}
\item{recommended_taxon_name}{The recommended taxon names.}
\item{recommended_TVK}{The UKSI taxon version keys associated with the taxa in recommended_taxon_name.}
}
}
\usage{
taxa_lookup
}
\description{
A lookup between the reccomended taxa as present in \code{RMAVIS::taxonomic_backbone} and associated synonymous taxon concepts present in the UKSI.
}
\details{
\code{taxa_lookup}
}
\keyword{datasets}
