% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acceptedSpecies}
\alias{acceptedSpecies}
\title{A dataset containing}
\format{
A data frame with 16500 rows and 3 columns, the definitions of which are:
\describe{
\item{}{}
}
}
\usage{
acceptedSpecies
}
\description{
\code{}
}
\keyword{datasets}
