% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_functions.R
\name{calc_rdiversity_metrics_meta}
\alias{calc_rdiversity_metrics_meta}
\title{...}
\usage{
calc_rdiversity_metrics_meta(rdiv_objects, q = 0)
}
\arguments{
\item{rdiv_objects}{...}

\item{q}{...}
}
\value{
...
}
\description{
...
}
\examples{
...
}
