% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_functions.R
\name{calc_rdiversity_metrics_subcom}
\alias{calc_rdiversity_metrics_subcom}
\title{Calculate diversity metrics for a subcommunity}
\usage{
calc_rdiversity_metrics_subcom(
  rdiv_objects,
  measures = c("alpha", "beta", "gamma"),
  metrics = c("naive", "taxonomic", "phylogenetic"),
  q = 0
)
}
\arguments{
\item{rdiv_objects}{A list of three objects: meta_naive, meta_tax, and meta_phylo_dist, produced using the function \code{RMAVIS::calc_rdiversity_objects}.}

\item{measures}{One or more of "alpha", "beta", and "gamma".}

\item{metrics}{One or more of "naive", "taxonomic", and "phylogenetic".}

\item{q}{A positive number representing a Hill-Number.}
}
\value{
A data frame containing the subcommunity partition diversity metrics for all combinations of specified measures and metrics
}
\description{
Calculate a set of subcommunity diversity measures and metrics using the objects
produced by the function \code{RMAVIS::calc_rdiversity_objects}.
}
\examples{
\dontrun{
rdiv_objs <- RMAVIS::calc_rdiversity_objects(plot_data = dplyr::filter(RMAVIS::example_data$`Parsonage Down`, Year == 1970 & Quadrat == "3N.1"), 
                                higher_taxa = dplyr::distinct(UKVegTB::taxonomic_backbone, taxon_name, Kingdom, Phylum, Class, Order, Family, Genus), 
                                phylo_tree = UKVegTB::phylo_tree, 
                                phylo_taxa_lookup = UKVegTB::phylo_taxa_lookup) |>
               RMAVIS::calc_rdiversity_metrics_subcom(measures = c("alpha", "beta", "gamma"), metrics = c("naive", "taxonomic", "phylogenetic"), q = 1)                                
}
}
