% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_functions.R
\name{calc_rdiversity_metrics_subcom}
\alias{calc_rdiversity_metrics_subcom}
\title{Calculate diversity metrics for a subcommunity}
\usage{
calc_rdiversity_metrics_subcom(
  rdiv_objects,
  measures = c("alpha", "beta", "gamma"),
  metrics = c("naive", "taxonomic", "phylogenetic"),
  q = 0
)
}
\arguments{
\item{rdiv_objects}{}

\item{measures}{One or more of "alpha", "beta", and "gamma".}

\item{metrics}{One or more of "naive", "taxonomic", and "phylogenetic".}

\item{q}{description}
}
\value{
...
}
\description{
...
}
\examples{
...
}
