% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_functions.R
\name{calc_rdiversity_objects}
\alias{calc_rdiversity_objects}
\title{...}
\usage{
calc_rdiversity_objects(
  plot_data,
  higher_taxa,
  phylo_tree,
  phylo_taxa_lookup,
  groups = c("Year", "Group", "Quadrat")
)
}
\arguments{
\item{plot_data}{...}

\item{higher_taxa}{...}

\item{phylo_tree}{...}

\item{phylo_taxa_lookup}{...}
}
\value{
...
}
\description{
...
}
\examples{
...
}
