% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntopic_table_functions.R
\name{composeSyntopicTables}
\alias{composeSyntopicTables}
\title{Compose a set of syntopic tables}
\usage{
composeSyntopicTables(
  surveyData,
  group_cols,
  species_col_name = "Species",
  plot_col_name = "Quadrat"
)
}
\arguments{
\item{surveyData}{A n-column dataframe containing atleast one ID columns, for example year and group.}

\item{group_cols}{A vector of columns to group the surveyData by and from which to compose syntopic tables}

\item{species_col_name}{The name of the species column}

\item{plot_col_name}{The name of the plot ID column}
}
\value{
A three column data frame containing the groups ID, species, and constancy.
}
\description{
Compose a set of syntopic/floristic/constancy tables for a
data frame containing plot ids, species names, and any number
of group columns which are concatenated into a ID column
}
