% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntopic_table_functions.R
\name{composeSyntopicTables}
\alias{composeSyntopicTables}
\title{Compose a set of syntopic tables}
\usage{
composeSyntopicTables(
  surveyData,
  group_cols,
  species_col_name = "Species",
  plot_col_name = "Quadrat",
  cover_col_name = "Cover",
  factor_constancy = TRUE,
  numeral_constancy = FALSE,
  remove_low_freq_taxa = NULL
)
}
\arguments{
\item{surveyData}{A n-column dataframe containing atleast one ID columns, for example year and group.}

\item{group_cols}{A vector of columns to group the surveyData by and from which to compose syntopic tables}

\item{species_col_name}{The name of the species column}

\item{plot_col_name}{The name of the plot ID column}

\item{cover_col_name}{The name of the cover/abundance cover}

\item{factor_constancy}{A boolean value (TRUE/FALSE) indicating whether the constancy column should be returned as a factor or not.}

\item{numeral_constancy}{If TRUE the numeric constancy classes (1, 2, 3, 4, 5) are transformed into roman numeral values (I, II, III, IV, V).}

\item{remove_low_freq_taxa}{If not NULL, a float greater than 0 and less than 1 indicating the minimum relative frequency of taxa to include in the floristic tables.}
}
\value{
A three column data frame containing the groups ID, species, and constancy.
}
\description{
Compose a set of syntopic/floristic/constancy tables for a
data frame containing plot ids, species names, and any number
of group columns which are concatenated into a ID column
}
\examples{
RMAVIS::composeSyntopicTables(surveyData = RMAVIS::example_data[["Newborough Warren"]], 
                              group_cols = c("Year", "Group"), 
                              species_col_name = "Species", 
                              plot_col_name = "Quadrat",
                              cover_col_name = "Cover",
                              factor_constancy = TRUE,
                              numeral_constancy = FALSE,
                              remove_low_freq_taxa = 0.05)
}
