% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hill_ellenberg}
\alias{hill_ellenberg}
\title{Hill-Ellenberg values for vascular plant and bryophyte taxa.}
\format{
A data frame with 3034 rows and 6 columns, the definitions of which are:
\describe{
\item{TVK}{The taxon, represented by the UKSI taxon version key.}
\item{F}{The Hill-Ellenberg moisture score}
\item{L}{The Hill-Ellenberg light score}
\item{N}{The Hill-Ellenberg nitrogen score}
\item{R}{The Hill-Ellenberg reaction score}
\item{S}{The Hill-Ellenberg salinity score}
}
}
\usage{
hill_ellenberg
}
\description{
A data frame containing the Hill-Ellenberg environmental indicator values for
vascular plant and bryophyte taxa.
Named 'hill_ellenberg' as additional trait and conservation status data will be
added in the future.
}
\details{
\code{hill_ellenberg}
}
\keyword{datasets}
