% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{master_data}
\alias{master_data}
\title{A dataset containing}
\format{
A data frame with 3096 rows and 9 columns, the definitions of which are:
\describe{
\item{}{}
}
}
\usage{
master_data
}
\description{
\code{}
}
\keyword{datasets}
