% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nvc_community_namesCodes}
\alias{nvc_community_namesCodes}
\title{A dataset containing}
\format{
A data frame with 883 rows and 3 columns, the definitions of which are:
\describe{
\item{}{}
}
}
\usage{
nvc_community_namesCodes
}
\description{
\code{}
}
\keyword{datasets}
