% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nvc_pquad_dca_noBryophytes_hulls}
\alias{nvc_pquad_dca_noBryophytes_hulls}
\title{A dataset containing}
\format{
A data frame with 25853 rows and 7 columns, the definitions of which are:
\describe{
\item{}{}
}
}
\usage{
nvc_pquad_dca_noBryophytes_hulls
}
\description{
\code{}
}
\keyword{datasets}
