% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nvc_pquads}
\alias{nvc_pquads}
\title{NVC pseudo-quadrats}
\format{
A data frame with 298098 rows and 3 columns, the definitions of which are:
\describe{
\item{nvc_code}{The NVC unit code.}
\item{psq_id}{The pseudo-quadrat ID}
\item{nvc_taxon_name}{The taxon name, see \code{RMAVIS::nvc_taxa_lookup}.}
}
}
\usage{
nvc_pquads
}
\description{
A dataset containing pseudo-quadrats for each NVC unit.
}
\details{
\code{nvc_pquads}
}
\keyword{datasets}
