% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_functions.R
\name{similarityCzekanowski}
\alias{similarityCzekanowski}
\title{Calculate the similarity between two vegetation plot datasets using the Czekanowski coefficient}
\usage{
similarityCzekanowski(
  samp_df,
  comp_df,
  samp_species_col,
  comp_species_col,
  samp_group_name,
  comp_group_name,
  samp_weight_name,
  comp_weight_name,
  downweight_threshold = 1,
  downweight_value = 0.1
)
}
\arguments{
\item{samp_df}{A data frame containing either sample vegetation plot data with cover values, or a syntopic/floristic table composed from sample vegetation plot data.}

\item{comp_df}{A data frame containing either}

\item{samp_species_col}{The name of the column in samp_df containing the species names.}

\item{comp_species_col}{The name of the column in comp_df containing the species names.}

\item{samp_group_name}{The name of the column in samp_df containing the group name.}

\item{comp_group_name}{The name of the column in comp_df containing the group name.}

\item{samp_weight_name}{The name of the column in samp_df containing the numeric weight values.}

\item{comp_weight_name}{The name of the column in comp_df containing the numeric weight values.}

\item{downweight_threshold}{A numeric value to replace with downweight_threshold.}

\item{downweight_value}{A numeric value to replace down_weight threshold with.}
}
\value{
A three column data frame containing the samp_group_name values, comp_group_name values, and similarity results (Similarity).
}
\description{
Calculate the Czekanowski's quantitative index/coefficient of similarity
using either a set of sample vegetation plots and a set of reference
vegetation plots; or a set of composed floristic/syntopic tables and a set of
references floristic/syntopic tables. Method follows Malloch (1998).
}
\examples{
RMAVIS::similarityCzekanowski(samp_df = RMAVIS::composeSyntopicTables(surveyData = RMAVIS::example_data[["Parsonage Down"]], 
                                                                      group_cols = c("Year", "Group"), 
                                                                      species_col_name = "Species", 
                                                                      plot_col_name = "Quadrat",
                                                                      numeric_constancy = TRUE), 
                              comp_df = RMAVIS::subset_nvcData(nvc_data = RMAVIS::nvc_floristic_tables, 
                                                               habitatRestriction = c("CG"), 
                                                               col_name = "nvc_code"), 
                              samp_species_col = "Species", 
                              comp_species_col = "nvc_taxon_name", 
                              samp_group_name = "ID", 
                              comp_group_name = "nvc_code",
                              samp_weight_name = "Constancy", 
                              comp_weight_name = "constancy",
                              downweight_threshold = 1, 
                              downweight_value = 0.1)
}
