% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_functions.R
\name{similarityJaccard}
\alias{similarityJaccard}
\title{Calculate the similarity between two vegetation plot datasets using the Jaccard coefficient}
\usage{
similarityJaccard(
  samp_df,
  comp_df,
  samp_species_col,
  comp_species_col = "nvc_taxon_name",
  samp_group_name = "ID",
  comp_group_name = "psq_id",
  comp_groupID_name = "nvc_code",
  remove_zero_matches = TRUE,
  average_comp = TRUE
)
}
\arguments{
\item{samp_df}{A data frame containing species presences (recorded in the samp_species_col column) by plot ID (recorded in the samp_species_col) column}

\item{comp_df}{A data frame containing species presences (recorded in the comp_species_col column) by plot ID (recorded in the comp_species_col) column with plot ID groups (recorded in the comp_groupID_column)}

\item{samp_species_col}{The name of the column in samp_df containing the species names.}

\item{comp_species_col}{The name of the column in comp_df containing the species names.}

\item{samp_group_name}{The name of the column in samp_df containing the group name.}

\item{comp_group_name}{The name of the column in comp_df containing the group name.}

\item{comp_groupID_name}{The name of the column containing the groups for each
comp_group_name value.}

\item{remove_zero_matches}{Exclude any results with a similarity of 0.}

\item{average_comp}{Take the mean average across the groups in comp_group_name by comp_groupID_name.}
}
\value{
A three column data frame containing the samp_group_name values, comp_group_name values or comp_groupID_name values, and similarity results (Similarity).
}
\description{
Calculate the Jaccard coefficient of similarity between two sets of vegetation
plot data. The Jaccard coefficient is a binary, presence/absence only
measure of similarity.
}
\examples{
RMAVIS::similarityJaccard(samp_df = RMAVIS::example_data[["Parsonage Down"]], 
                          comp_df = RMAVIS::subset_nvcData(nvc_data = RMAVIS::nvc_pquads, 
                                                           habitatRestriction = c("CG"), 
                                                           col_name = "psq_id"),
                          samp_species_col = "Species", 
                          comp_species_col = "nvc_taxon_name",
                          samp_group_name = "Quadrat", 
                          comp_group_name = "psq_id",
                          comp_groupID_name = "nvc_code", 
                          remove_zero_matches = TRUE, 
                          average_comp = TRUE)
}
